requirements.txt
streamlit>=1.35.0
langchain>=1.2.0
langchain-classic>=1.0.0
langchain-core>=1.2.0
langchain-community>=0.4.0
langchain-openai>=1.1.0
langchain-text-splitters>=1.1.0
chromadb>=0.5.0
openai>=2.0.0
pypdf>=4.0.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
